\name{defaultSessId}
\Rdversion{1.1}
\alias{defaultSessId}
\docType{data}
\title{
defaultSessId
}
\description{
defaultSessId is a public variable that holds an empty string.
When session starts, it's value set to RSESSID cookie. If RSESSID cookie has an empty value,
randomly generated session id is used for defaultSessId.
}

\keyword{datasets}
