\name{response.println}
\Rdversion{1.1}
\alias{response.println}

\title{
response.println
}
\description{
A wrapper method for response.print(). This function puts a return character after all of the given parameters sent 
to clients.
}
\usage{
response.println(...)
}

\arguments{
  \item{\dots}{
Parameters that seperated with commas
}
}

\author{
Mehmet Hakan Satman
}



\seealso{
\link{response.print}
}
\examples{
	response.println("Hello World!")
}
